backend/: move newMultipleObjectsResponse to response.go
